Revert "make throughput unsigned"
This reverts commit 82713a8ac14e0735ad4e828452c56e7cde4809e1.

+++ JHA/jackpotcoin.cu
@@ -96,8 +96,8 @@ extern "C" int scanhash_jackpot(int thr_id, uint32_t *pdata,
 	if (opt_benchmark)
 		((uint32_t*)ptarget)[7] = 0x000f;
 
-	uint32_t throughput = opt_work_size ? opt_work_size : (1 << 20); // 256*4096
-	throughput = min(throughput, max_nonce - first_nonce);
+	int throughput = opt_work_size ? opt_work_size : (1 << 20); // 256*4096
+	throughput = min(throughput, (int)(max_nonce - first_nonce));
 
 	if (!init[thr_id])
 	{
